osquery: Move getWindowsErrorDescription to errno
errno.h is a more sensible location than filepath.h
