+++ osquery/utils/system/errno.h
@@ -9,10 +9,16 @@
 #pragma once
 
 #include <string>
+#include <osquery/utils/status/status.h>
 
 namespace osquery {
 
 /// Returns a C++ string explaining the errnum
 std::string platformStrerr(int errnum);
+
+#ifdef WIN32
+/// Converts a Windows error (winerror.h/GetLastError()) to a string
+Status getWindowsErrorDescription(std::string& error_message, DWORD error_id);
+#endif
 
 }
